Working on engine import/pkg improvements to make content dev simpler
