+++ package.json
@@ -11,6 +11,12 @@
         "start:update": "node --max-old-space-size=1024 dist/game-engine/update-server.js",
         "start:infra": "concurrently \"npm run start:update\" \"npm run start:login\"",
         "start:standalone": "concurrently \"npm run start:infra\" \"npm run start:game\"",
+        "game": "npm run start:game",
+        "game:dev": "npm run start:game",
+        "login": "npm run start:login",
+        "update": "npm run start:update",
+        "infra": "npm run start:infra",
+        "standalone": "npm run start:standalone",
         "lint": "eslint --ext .ts src",
         "lint:fix": "eslint --ext .ts src --fix",
         "build:watch": "babel ./src --out-dir dist --extensions \".ts,.tsx,.js\" --source-maps --watch",
