+++ src/game-engine/plugins/index.ts
@@ -1,0 +1,1 @@
+export * from './content-plugin';
