+++ src/game-engine/world/action/item-interaction.action.ts
@@ -3,7 +3,7 @@ import { ActionHook, getActionHooks } from '@engine/world/action/hooks';
 import { ItemDetails } from '@engine/config/item-config';
 import { findItem } from '@engine/config';
 import { numberHookFilter, stringHookFilter, questHookFilter } from '@engine/world/action/hooks/hook-filters';
-import { ActionPipe, RunnableHooks } from '@engine/world/action/index';
+import { ActionPipe, RunnableHooks } from '@engine/world/action/action-pipeline';
 
 
 /**
