+++ src/game-engine/world/action/item-on-object.action.ts
@@ -3,7 +3,7 @@ import { Position } from '@engine/world/position';
 import { ActionHook, getActionHooks } from '@engine/world/action/hooks';
 import { Item } from '@engine/world/items/item';
 import { advancedNumberHookFilter, questHookFilter } from '@engine/world/action/hooks/hook-filters';
-import { ActionPipe, RunnableHooks } from '@engine/world/action/index';
+import { ActionPipe, RunnableHooks } from '@engine/world/action/action-pipeline';
 import { LandscapeObject, ObjectConfig } from '@runejs/filestore';
 import { playerWalkTo } from '@engine/game-server';
