+++ src/game-engine/world/action/item-swap.action.ts
@@ -1,7 +1,7 @@
 import { Player } from '../actor/player/player';
 import { ActionHook, getActionHooks } from '@engine/world/action/hooks';
 import { numberHookFilter } from '@engine/world/action/hooks/hook-filters';
-import { ActionPipe, RunnableHooks } from '@engine/world/action/index';
+import { ActionPipe, RunnableHooks } from '@engine/world/action/action-pipeline';
 
 
 /**
