+++ src/plugins/npcs/lumbridge/bob.plugin.ts
@@ -1,19 +1,19 @@
-import { npcInteractionActionHandler } from '@engine/world/action/npc-interaction.action';
+import { ContentPlugin } from '@engine/plugins';
+import { NpcInteractionActionHook } from '@engine/world/action';
 import { findShop } from '@engine/config';
 
 
-const action: npcInteractionActionHandler = ({ player }) =>
-    findShop('rs:lumbridge_bobs_axes')?.open(player);
-
-export default {
-    pluginId: 'rs:bob',
-    hooks: [
-        {
-            type: 'npc_interaction',
-            npcs: 'rs:lumbridge_bob',
-            options: 'trade',
-            walkTo: true,
-            handler: action
-        }
-    ]
+const bobHook: NpcInteractionActionHook = {
+    type: 'npc_interaction',
+    npcs: 'rs:lumbridge_bob',
+    options: 'trade',
+    walkTo: true,
+    handler: ({ player }) => findShop('rs:lumbridge_bobs_axes')?.open(player);
 };
+
+const bobPlugin: ContentPlugin = {
+    pluginId: 'rs:bob',
+    hooks: [ bobHook ]
+}
+
+export default bobPlugin;
